Prepare for v1.0.0
