+++ docs/classes/LazyCollection.Stream.html
@@ -114,7 +114,7 @@
                         <ul class="dropdown-menu">
                             <li>
                                 <a href="../reports/errors.html">
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    <i class="icon-list-alt"></i>&#160;Errors <span class="label label-info pull-right">15</span>
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    <i class="icon-list-alt"></i>&#160;Errors <span class="label label-info pull-right">14</span>
                                 </a>
                             </li>
                             <li>
@@ -144,28 +144,28 @@
                                 <div class="accordion" style="margin-bottom: 0">
         <div class="accordion-group">
             <div class="accordion-heading">
-                                    <a class="accordion-toggle " data-toggle="collapse" data-target="#namespace-514076490"></a>
+                                    <a class="accordion-toggle " data-toggle="collapse" data-target="#namespace-694879703"></a>
                                 <a href="../namespaces/default.html" style="margin-left: 30px; padding-left: 0">\</a>
             </div>
-            <div id="namespace-514076490" class="accordion-body collapse in">
+            <div id="namespace-694879703" class="accordion-body collapse in">
                 <div class="accordion-inner">
 
                                                     <div class="accordion" style="margin-bottom: 0">
         <div class="accordion-group">
             <div class="accordion-heading">
-                                    <a class="accordion-toggle collapsed" data-toggle="collapse" data-target="#namespace-2112105226"></a>
+                                    <a class="accordion-toggle collapsed" data-toggle="collapse" data-target="#namespace-1130486287"></a>
                                 <a href="../namespaces/LazyCollection.html" style="margin-left: 30px; padding-left: 0">LazyCollection</a>
             </div>
-            <div id="namespace-2112105226" class="accordion-body collapse ">
+            <div id="namespace-1130486287" class="accordion-body collapse ">
                 <div class="accordion-inner">
 
                                                     <div class="accordion" style="margin-bottom: 0">
         <div class="accordion-group">
             <div class="accordion-heading">
-                                    <a class="accordion-toggle collapsed" data-toggle="collapse" data-target="#namespace-1005185837"></a>
+                                    <a class="accordion-toggle collapsed" data-toggle="collapse" data-target="#namespace-2133485690"></a>
                                 <a href="../namespaces/LazyCollection.Exceptions.html" style="margin-left: 30px; padding-left: 0">Exceptions</a>
             </div>
-            <div id="namespace-1005185837" class="accordion-body collapse ">
+            <div id="namespace-2133485690" class="accordion-body collapse ">
                 <div class="accordion-inner">
 
                     
@@ -320,7 +320,6 @@
                             <section class="span4">
                                                                     <a href="../classes/LazyCollection.Stream.html#method___construct" class="">__construct()</a><br />
                                                                     <a href="../classes/LazyCollection.Stream.html#method_pipe" class="">pipe()</a><br />
-                                                                    <a href="../classes/LazyCollection.Stream.html#method_nullInstance" class="">nullInstance()</a><br />
                                                                     <a href="../classes/LazyCollection.Stream.html#method_hasMethod" class="">hasMethod()</a><br />
                                                                     <a href="../classes/LazyCollection.Stream.html#method_hasFactory" class="">hasFactory()</a><br />
                                                             </section>
@@ -3715,35 +3714,6 @@
 
                                     <div class="row-fluid">
         <div class="span8 content class">
-            <a id="method_nullInstance" name="method_nullInstance" class="anchor"></a>
-            <article class="method">
-                <h3 class="protected ">nullInstance()</h3>
-                <a href="#source-view" role="button" class="pull-right btn" data-toggle="modal" style="font-size: 1.1em; padding: 9px 14px"><i class="icon-code"></i></a>
-                <pre class="signature" style="margin-right: 54px;">nullInstance() </pre>
-                <p><em></em></p>
-                
-
-                
-                
-                
-                            </article>
-        </div>
-        <aside class="span4 detailsbar">
-            <h1><i class="icon-arrow-down"></i></h1>
-                            <span class="label label-info">static</span>
-                                                            <dl>
-                                                    <dt>Inherited from</dt>
-                    <dd><a href="../classes/LazyCollection.Stream.html"><div class="path-wrapper">\LazyCollection\Stream</div></a></dd>
-                                                            </dl>
-            <h2>Tags</h2>
-            <table class="table table-condensed">
-                                    <tr><td colspan="2"><em>None found</em></td></tr>
-                            </table>
-        </aside>
-    </div>
-
-                                    <div class="row-fluid">
-        <div class="span8 content class">
             <a id="method_hasMethod" name="method_hasMethod" class="anchor"></a>
             <article class="method">
                 <h3 class="protected ">hasMethod()</h3>
@@ -3874,7 +3844,7 @@
                 <section class="span10 offset1">
                     <hr />
                     Documentation is powered by <a href="http://www.phpdoc.org/">phpDocumentor </a> and authored
-                    on October 9th, 2020 at 17:04.
+                    on October 9th, 2020 at 20:27.
                 </section>
             </section>
         </section>
