+++ docs/reports/errors.html
@@ -67,7 +67,7 @@
                         <ul class="dropdown-menu">
                             <li>
                                 <a href="../reports/errors.html">
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    <i class="icon-list-alt"></i>&#160;Errors <span class="label label-info pull-right">15</span>
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    <i class="icon-list-alt"></i>&#160;Errors <span class="label label-info pull-right">14</span>
                                 </a>
                             </li>
                             <li>
@@ -122,7 +122,7 @@
                         <h3>
                             <i class="icon-file"></i>
                             Stream.php
-                            <small style="float: right;padding-right: 10px;">3</small>
+                            <small style="float: right;padding-right: 10px;">2</small>
                         </h3>
                         <div>
                             <table class="table markers table-bordered">
@@ -141,12 +141,7 @@
                                     </tr>
                                                                     <tr>
                                         <td>error</td>
-                                        <td>191</td>
-                                        <td>No summary for method nullInstance()</td>
-                                    </tr>
-                                                                    <tr>
-                                        <td>error</td>
-                                        <td>260</td>
+                                        <td>251</td>
                                         <td>No summary for method getIterator()</td>
                                     </tr>
                                                             </tbody>
@@ -442,7 +437,7 @@
                 <section class="span10 offset1">
                     <hr />
                     Documentation is powered by <a href="http://www.phpdoc.org/">phpDocumentor </a> and authored
-                    on October 9th, 2020 at 17:04.
+                    on October 9th, 2020 at 20:27.
                 </section>
             </section>
         </section>
