+++ tests/Methods/Operators/Limiters/TakeTest.php
@@ -1,13 +1,13 @@
 <?php
 
 
-namespace LazyCollection\Tests\Methods\StatefulOperators\Limiters;
+namespace LazyCollection\Tests\Methods\Operators\Limiters;
 
 
 use LazyCollection\Stream;
 use LazyCollection\Tests\PHPUnit;
 
-class SkipTest extends PHPUnit
+class TakeTest extends PHPUnit
 {
 	/******************************************************************************************************************\
 	 * HELPERS
@@ -17,9 +17,9 @@ class SkipTest extends PHPUnit
 	 * @param int $maxAmount
 	 * @return array
 	 */
-	public function skip(iterable $it, int $maxAmount){
+	public function take(iterable $it, int $maxAmount){
 		return Stream::fromIterable($it)
-			->skip($maxAmount)
+			->take($maxAmount)
 			->toArray();
 	}
 
@@ -30,15 +30,15 @@ class SkipTest extends PHPUnit
 	\******************************************************************************************************************/
 	/**
 	 * @test
-	 * @cover \LazyCollection\Stream::skip
+	 * @cover \LazyCollection\Stream::take
 	 * @dataProvider provideTakeData
 	 *
 	 * @param iterable $input
 	 * @param int $maxAmount
 	 * @param iterable $expected
 	 */
-	public function makeSureItSkipsUpToTheMaxAmount(iterable $input, int $maxAmount, iterable $expected){
-		$result = $this->skip($input, $maxAmount);
+	public function makeSureItTakesUpToTheMaxAmount(iterable $input, int $maxAmount, iterable $expected){
+		$result = $this->take($input, $maxAmount);
 		$this->assertEquals($expected, $result);
 	}
 
@@ -50,9 +50,9 @@ class SkipTest extends PHPUnit
 	public function provideTakeData(){
 		return [
 			[
-				[1, 2, 3, 4],
+				[1,2,3,4],
 				2,
-				[3, 4],
+				[1,2],
 			],
 			[
 				[],
@@ -62,12 +62,12 @@ class SkipTest extends PHPUnit
 			[
 				[1],
 				3,
-				[],
+				[1],
 			],
 			[
 				[1,2,3,4],
 				4,
-				[],
+				[1,2,3,4],
 			],
 		];
 	}
