+++ tests/Methods/Operators/Limiters/TakeUntilTest.php
@@ -1,25 +1,25 @@
 <?php
 
 
-namespace LazyCollection\Tests\Methods\StatefulOperators\Limiters;
+namespace LazyCollection\Tests\Methods\Operators\Limiters;
 
 
 use LazyCollection\Stream;
 use LazyCollection\Tests\PHPUnit;
 
-class TakeTest extends PHPUnit
+class TakeUntilTest extends PHPUnit
 {
 	/******************************************************************************************************************\
 	 * HELPERS
 	\******************************************************************************************************************/
 	/**
 	 * @param iterable $it
-	 * @param int $maxAmount
+	 * @param callable $predicate
 	 * @return array
 	 */
-	public function take(iterable $it, int $maxAmount){
+	public function takeUntil(iterable $it, callable $predicate){
 		return Stream::fromIterable($it)
-			->take($maxAmount)
+			->takeUntil($predicate)
 			->toArray();
 	}
 
@@ -30,15 +30,15 @@ class TakeTest extends PHPUnit
 	\******************************************************************************************************************/
 	/**
 	 * @test
-	 * @cover \LazyCollection\Stream::take
-	 * @dataProvider provideTakeData
+	 * @cover \LazyCollection\Stream::takeUntil
+	 * @dataProvider provideTakeUntilData
 	 *
 	 * @param iterable $input
-	 * @param int $maxAmount
+	 * @param callable $predicate
 	 * @param iterable $expected
 	 */
-	public function makeSureItTakesUpToTheMaxAmount(iterable $input, int $maxAmount, iterable $expected){
-		$result = $this->take($input, $maxAmount);
+	public function properlyTakesElementUntilThePredicateIsFulfilled(iterable $input, callable $predicate, iterable $expected){
+		$result = $this->takeUntil($input, $predicate);
 		$this->assertEquals($expected, $result);
 	}
 
@@ -47,27 +47,17 @@ class TakeTest extends PHPUnit
 	/******************************************************************************************************************\
 	 * TEST PROVIDERS
 	\******************************************************************************************************************/
-	public function provideTakeData(){
+	public function provideTakeUntilData(){
 		return [
 			[
-				[1,2,3,4],
-				2,
-				[1,2],
+				[1, 2, 3, 4, 5, 6],
+				function($e){ return $e >= 5; },
+				[1, 2, 3, 4],
 			],
 			[
-				[],
-				2,
-				[],
-			],
-			[
-				[1],
-				3,
-				[1],
-			],
-			[
-				[1,2,3,4],
-				4,
-				[1,2,3,4],
+				[2, 4, 3, 4],
+				function($e){ return $e % 2 !== 0; },
+				[2, 4],
 			],
 		];
 	}
