+++ tests/Methods/Operators/Limiters/TakeWhileTest.php
@@ -1,13 +1,13 @@
 <?php
 
 
-namespace LazyCollection\Tests\Methods\StatefulOperators\Limiters;
+namespace LazyCollection\Tests\Methods\Operators\Limiters;
 
 
 use LazyCollection\Stream;
 use LazyCollection\Tests\PHPUnit;
 
-class TakeUntilTest extends PHPUnit
+class TakeWhileTest extends PHPUnit
 {
 	/******************************************************************************************************************\
 	 * HELPERS
@@ -17,9 +17,9 @@ class TakeUntilTest extends PHPUnit
 	 * @param callable $predicate
 	 * @return array
 	 */
-	public function takeUntil(iterable $it, callable $predicate){
+	public function takeWhile(iterable $it, callable $predicate){
 		return Stream::fromIterable($it)
-			->takeUntil($predicate)
+			->takeWhile($predicate)
 			->toArray();
 	}
 
@@ -30,15 +30,15 @@ class TakeUntilTest extends PHPUnit
 	\******************************************************************************************************************/
 	/**
 	 * @test
-	 * @cover \LazyCollection\Stream::takeUntil
-	 * @dataProvider provideTakeUntilData
+	 * @cover \LazyCollection\Stream::takeWhile
+	 * @dataProvider provideTakeWhileData
 	 *
 	 * @param iterable $input
 	 * @param callable $predicate
 	 * @param iterable $expected
 	 */
-	public function properlyTakesElementUntilThePredicateIsFulfilled(iterable $input, callable $predicate, iterable $expected){
-		$result = $this->takeUntil($input, $predicate);
+	public function properlyTakesElementWhileThePredicateIsFulfilled(iterable $input, callable $predicate, iterable $expected){
+		$result = $this->takeWhile($input, $predicate);
 		$this->assertEquals($expected, $result);
 	}
 
@@ -47,16 +47,16 @@ class TakeUntilTest extends PHPUnit
 	/******************************************************************************************************************\
 	 * TEST PROVIDERS
 	\******************************************************************************************************************/
-	public function provideTakeUntilData(){
+	public function provideTakeWhileData(){
 		return [
 			[
 				[1, 2, 3, 4, 5, 6],
-				function($e){ return $e >= 5; },
+				function($e){ return $e < 5; },
 				[1, 2, 3, 4],
 			],
 			[
 				[2, 4, 3, 4],
-				function($e){ return $e % 2 !== 0; },
+				function($e){ return $e % 2 === 0; },
 				[2, 4],
 			],
 		];
